+++ src/test/kotlin/com/zepben/evolve/services/network/tracing/phases/PhaseLogger.kt
@@ -10,10 +10,11 @@ package com.zepben.evolve.services.network.tracing.phases
 import com.zepben.evolve.cim.iec61970.base.core.ConductingEquipment
 import com.zepben.evolve.services.common.extensions.typeNameAndMRID
 import com.zepben.evolve.services.network.tracing.Tracing
+import com.zepben.evolve.services.network.tracing.connectivity.ConductingEquipmentStep
 import org.slf4j.LoggerFactory
 
 // Logs all the phases of assets, terminals and nominal phases. Useful for debugging.
-internal class PhaseLogger private constructor(asset: ConductingEquipment) : (ConductingEquipment, Boolean?) -> Unit {
+internal class PhaseLogger private constructor(asset: ConductingEquipment) : (ConductingEquipmentStep, Boolean?) -> Unit {
 
     private val b: StringBuilder = StringBuilder()
         .append("\n###############################")
@@ -21,9 +22,9 @@ internal class PhaseLogger private constructor(asset: ConductingEquipment) : (Co
         .append("\n")
         .append("\n")
 
-    override fun invoke(a: ConductingEquipment, isStopping: Boolean?) {
-        a.terminals.forEach { t ->
-            b.append("${a.mRID}-T${t.sequenceNumber}: ")
+    override fun invoke(a: ConductingEquipmentStep, isStopping: Boolean?) {
+        a.conductingEquipment.terminals.forEach { t ->
+            b.append("${a.conductingEquipment.mRID}-T${t.sequenceNumber}: ")
 
             t.phases.singlePhases.forEach { phase ->
                 val nps = t.normalPhases[phase]
@@ -60,8 +61,7 @@ internal class PhaseLogger private constructor(asset: ConductingEquipment) : (Co
             assets.forEach { asset ->
                 val pl = PhaseLogger(asset!!)
 
-                Tracing.connectedEquipmentTrace()
-                    .addStepAction(pl)
+                Tracing.connectedEquipmentTrace().apply { addStepAction(pl) }
                     .run(asset)
 
                 pl.log()